requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
